requirements.txt
numpy
scikit-learn
scipy
opencv
torch
torchvision
torchaudio

absl-py
coloredlogs
wandb
seaborn
accelerate
diffusers["torch"]
blobfile
einops
path.py
matplotlib
ml_collections
tqdm
tabulate
ipdb
labml_nn